+++ backend/src/main/java/com/woowacourse/kkogkkog/infrastructure/dto/PushAlarmRequest.java
@@ -1,4 +1,4 @@
-package com.woowacourse.kkogkkog.infrastructure;
+package com.woowacourse.kkogkkog.infrastructure.dto;
 
 import lombok.Getter;
